% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{elgreco_palette}
\alias{elgreco_palette}
\title{An El Greco palette generator}
\usage{
elgreco_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{AndrewFrancis1}, \code{AndrewFrancis2},  \code{Annunciation1},
\code{Annunciation2}, \code{Caridad1},  \code{Caridad2}, \code{Disrobing1},
\code{Disrobing2},  \code{Immaculate1} , \code{Immaculate2} ,
\code{Laocoon1}, \code{Laocoon2}, \code{Marriage1}, \code{Marriage2},
\code{Shepherds1}, \code{Luke1}, \code{Luke2}, \code{Toledo1},
\code{Toledo2}, \code{John1}, \code{John2}}

\item{n}{Number of colors desired. All color
schemes are derived from El Greco paintings.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
Color palettes from El Greco paintings.
}
\examples{
elgreco_palette("Laocoon1")
elgreco_palette("Shepherds1")
elgreco_palette("Toledo1")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- elgreco_palette(21, name = "Toledo1", type = "continuous")
}
\keyword{colors}
