% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{elgreco_palettes}
\alias{elgreco_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 21.
}
\usage{
elgreco_palettes
}
\description{
Use \code{\link{elgreco_palette}} to construct palettes of desired length.
}
\keyword{datasets}
